requirements.production.txt
# Production Requirements for PipLinePro
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
psycopg2-pool==1.1
alembic==1.12.0

# Redis and Caching
redis==5.0.1
Flask-Caching==2.1.0
Flask-Session==0.5.0

# HTTP and API
requests==2.31.0
urllib3==2.0.4
httpx==0.24.1

# Security
cryptography==41.0.4
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.1.0
filetype==1.2.0

# Production WSGI Server
gunicorn==21.2.0
gevent==23.7.0

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.1.0
python-json-logger==2.0.7
psutil==5.9.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# Exchange Rate Dependencies
yfinance==0.2.65

# Visualization and Analytics
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Email (optional)
Flask-Mail==0.9.1

# File Processing
Pillow==10.0.1
python-magic==0.4.27

# Backup and Storage
boto3==1.28.57
azure-storage-blob==12.19.0

# Development and Testing (for CI/CD)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
bandit==1.7.5

# Performance
uvloop==0.17.0
ujson==5.8.0

# Health Checks
healthcheck==1.3.3

# Rate Limiting
slowapi==0.1.9

# Database Connection Pooling
sqlalchemy-pool==0.1.0

# Compression
gzip-stream==0.1.0
Flask-Compress==1.18

# Configuration Management
dynaconf==3.2.0

# Metrics and Analytics
statsd==4.0.1

# Background Tasks
celery==5.3.4
kombu==5.3.2
flower==2.0.1
schedule==1.2.0

# WebSocket Support
flask-socketio==5.3.6
eventlet==0.33.3

# API Documentation
flask-restx==1.1.0

# Validation
marshmallow==3.20.1
webargs==8.3.0

# Internationalization
Flask-Babel==4.0.0
Babel==2.14.0

# File Upload
Flask-Uploads==0.2.1

# Error Handling
sentry-sdk[flask]==1.32.0

# Production Utilities
supervisor==4.2.5

# Database Recovery and Maintenance
# Note: sqlite3 is included in Python standard library
# All database recovery functionality uses standard library modules
# No additional dependencies required for database recovery system
